fix: resolve lint errors, add concurrency, and fix wrk2 build
- Fix all golangci-lint v2.8.0 errors (errcheck, unconvert) in theoretical servers
- Add 'libssl-dev' to uerdata.tftpl for wrk2 openssl dependency
- Add concurrency groups to workflows to enforce sequential runs per ref

+++ servers/theoretical/iouring/http2.go
@@ -197,7 +197,7 @@ func (s *HTTP2Server) submitMultishotRecv(fd int) {
 	sqe.OpcodeFlags = IORING_RECV_MULTISHOT | (bufferGroup << 16)
 	sqe.UserData = uint64(fd)
 
-	s.sqArray[idx] = uint32(idx)
+	s.sqArray[idx] = idx
 	*s.sqTail = tail + 1
 }
 
@@ -212,7 +212,7 @@ func (s *HTTP2Server) submitSend(fd int, data []byte) {
 	sqe.Len = uint32(len(data))
 	sqe.UserData = uint64(fd) | (1 << 32)
 
-	s.sqArray[idx] = uint32(idx)
+	s.sqArray[idx] = idx
 	*s.sqTail = tail + 1
 }
 
@@ -269,7 +269,7 @@ func (s *HTTP2Server) reprovideBuffer(bufIdx int) {
 	sqe.Off = uint64(bufIdx)
 	sqe.OpcodeFlags = bufferGroup
 
-	s.sqArray[idx] = uint32(idx)
+	s.sqArray[idx] = idx
 	*s.sqTail = tail + 1
 }
